SEQUOIADBMAINSTREAM-2683 sdbdmsdump支持固定集合;SEQUOIADBMAINSTREAM-2680 增加了mb中的字段后的升级处理;编译告警处理
git-svn-id: http://192.168.20.11/sequoiadb/trunk@30379 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/dms/dmsStorageDataCapped.cpp
@@ -473,7 +473,7 @@ namespace engine
          }
 
          // If free space in current working extent is not enough, get another.
-         if ( workExtInfo->_freeSpace < (INT32)size )
+         if ( workExtInfo->_freeSpace < size )
          {
             // Before allocationg another extent, check size limitation.
             if ( _sizeExceedLimit( context, DMS_CAP_EXTENT_SZ ) )
@@ -495,7 +495,7 @@ namespace engine
             // After recycle, check once again. For when recycling the woking
             // extent, it will be reset and used again. In that case, no need to
             // allocate extent again.
-            if ( workExtInfo->_freeSpace < (INT32)size )
+            if ( workExtInfo->_freeSpace < size )
             {
                rc = _allocateExtent( context, DMS_CAP_EXTENT_PAGE_NUM, TRUE,
                                      FALSE, &extID ) ;
@@ -635,7 +635,7 @@ namespace engine
                }
             }
             // If free space in current working extent is not enough, get another.
-            if ( workExtInfo->_freeSpace < (INT32)size )
+            if ( workExtInfo->_freeSpace < size )
             {
                // Before allocationg another extent, check size limitation.
                if ( _sizeExceedLimit( context, DMS_CAP_EXTENT_SZ ) )
@@ -657,7 +657,7 @@ namespace engine
                // After recycle, check once again. For when recycling the woking
                // extent, it will be reset and used again. In that case, no need to
                // allocate extent again.
-               if ( workExtInfo->_freeSpace < (INT32)size )
+               if ( workExtInfo->_freeSpace < size )
                {
                   rc = _allocateExtent( context, DMS_CAP_EXTENT_PAGE_NUM, TRUE,
                                         FALSE, &extID ) ;
@@ -983,7 +983,7 @@ namespace engine
                maxSize = ele.numberLong() ;
 
                if ( maxSize < DMS_CAP_CL_MIN_SZ ||
-                    maxSize > DMS_MAX_SZ( pageSize() ) )
+                    (UINT64)maxSize > DMS_MAX_SZ( pageSize() ) )
                {
                   PD_LOG( PDERROR, "Invalid size[%lld] of capped collection",
                           maxSize ) ;
